成功用4090和本地微调了base模型，本地的效果非常差，后续将都使用4090

+++ my_model.py
@@ -320,7 +320,7 @@ class LLMIEForCausalLM(LLMIEPreTrainedModel, GenerationMixin):
             **kwargs,
         )
         hidden_states = outputs[0] if not return_dict else outputs.last_hidden_state
-        logits = self.lm_head(hidden_states).float()
+        logits = self.lm_head(hidden_states)  # 保持 bf16/fp16，由 Trainer autocast 管理精度
 
         loss = None
         if labels is not None:
